fixup! replace autobahn&crossbar by gRPC
The pump_task might have already exited due to a connection error (e.g.
when the coordinator is unreachable). In that case, self.out_queue.join()
blocks forever because nobody removes items from the queue anymore.

Do not block on non-empty queue. Instead join the queue with a timeout.
If the timeout happens, check if the pump_task is still running. If not,
return early. There is no point in creating the poll task in this case.
It only obscures the error by logging resources after the connection
error already happened.

Signed-off-by: Bastian Krause <bst@pengutronix.de>